task5/client/requirements.txt
zeroc-ice>=3.6.4
